+++ gcmt/gcmtools.py
@@ -14,7 +14,7 @@ import gcmt.core.writer as wrt
 import gcmt.utils.gcm_plotting as gcmplt
 import gcmt.utils.manipulations as mani
 import gcmt.utils.read_and_write as raw
-from gcmt.gcmDatasetCollection import GCMDatasetCollection
+from gcmt.gcm_dataset_collection import GCMDatasetCollection
 from gcmt.core.units import ALLOWED_PUNITS, ALLOWED_TIMEUNITS
 from gcmt.utils.passport import is_the_data_basic
 from gcmt.utils.interface import  PrtInterface
